fix: 분석 로직 수정

+++ backend/src/main/java/com/buildsync/dto/analysis/MonthlyMaterialCostProjection.java
@@ -4,8 +4,8 @@ public interface MonthlyMaterialCostProjection {
 
     String getMonth();
 
-    Integer getTotalQuantity();
+    Long getTotalQuantity();
 
-    Integer getTotalMaterialCost();
+    Long getTotalMaterialCost();
 
 }
